readyup|refactor: Move the nearest-word matcher into `portable`
`findNearestWord` is a string matcher over a caller-supplied candidate list with no rdy domain in it, which is what `src/portable/` holds. It sat in `src/bin/` because that is where `route.ts` was; with `help/helpCommand.ts` now a second consumer, the module lived inside one of its callers' directories.

The file and its suite move unchanged, and the two importers name the new path.
