+++ packages/readyup/src/portable/findNearestWord.ts
